Moved ingest pipeline related code to separate module
