+++ tests/unit_test/opensearch/bigquery_to_opensearch_pipeline_test.py
@@ -23,13 +23,8 @@ from data_pipeline.opensearch.bigquery_to_opensearch_config import (
 )
 import data_pipeline.opensearch.bigquery_to_opensearch_pipeline as test_module
 from data_pipeline.opensearch.bigquery_to_opensearch_pipeline import (
-    assert_opensearch_ingest_pipeline_test_actual_documents_match_expected,
-    get_opensearch_ingest_pipeline_test_actual_documents_from_simulate_response,
-    get_opensearch_ingest_pipeline_tests_body,
-    run_opensearch_ingest_pipeline_tests,
     setup_opensearch_and_load_documents_into_opensearch,
     create_or_update_opensearch_index,
-    create_or_update_opensearch_ingest_pipeline,
     fetch_documents_from_bigquery_and_load_into_opensearch,
     fetch_documents_from_bigquery_and_load_into_opensearch_from_config_list,
     get_opensearch_client,
@@ -181,25 +176,9 @@ def _get_opensearch_client_mock(opensearch_client_mock: MagicMock) -> Iterator[M
         yield mock
 
 
-@pytest.fixture(name='assert_opensearch_ingest_pipeline_test_actual_documents_match_expected_mock')
-def _assert_opensearch_ingest_pipeline_test_actual_documents_match_expected_mock(
-) -> Iterator[MagicMock]:
-    with patch.object(
-        test_module,
-        'assert_opensearch_ingest_pipeline_test_actual_documents_match_expected'
-    ) as mock:
-        yield mock
-
-
-@pytest.fixture(name='run_opensearch_ingest_pipeline_tests_mock')
-def _run_opensearch_ingest_pipeline_tests_mock() -> Iterator[MagicMock]:
-    with patch.object(test_module, 'run_opensearch_ingest_pipeline_tests') as mock:
-        yield mock
-
-
-@pytest.fixture(name='create_or_update_opensearch_ingest_pipeline_mock')
+@pytest.fixture(name='create_or_update_opensearch_ingest_pipelines_mock')
 def _create_or_update_opensearch_ingest_pipeline_mock() -> Iterator[MagicMock]:
-    with patch.object(test_module, 'create_or_update_opensearch_ingest_pipeline') as mock:
+    with patch.object(test_module, 'create_or_update_opensearch_ingest_pipelines') as mock:
         yield mock
 
 
@@ -407,132 +386,6 @@ class TestGetOpenSearchClient:
         assert client == opensearch_class_mock.return_value
 
 
-class TestGetOpenSearchIngestPipelineTestsBody:
-    def test_should_parse_pipeline_definition_as_json(self):
-        body = get_opensearch_ingest_pipeline_tests_body(
-            ingest_pipeline_config=dataclasses.replace(
-                OPENSEARCH_INGEST_PIPELINE_CONFIG_1,
-                tests=[OPENSEARCH_INGEST_PIPELINE_TEST_CONFIG_1]
-            )
-        )
-        assert body['pipeline'] == json.loads(OPENSEARCH_INGEST_PIPELINE_CONFIG_1.definition)
-
-    def test_should_include_input_documents_as_source(self):
-        body = get_opensearch_ingest_pipeline_tests_body(
-            ingest_pipeline_config=dataclasses.replace(
-                OPENSEARCH_INGEST_PIPELINE_CONFIG_1,
-                tests=[OPENSEARCH_INGEST_PIPELINE_TEST_CONFIG_1]
-            )
-        )
-        assert len(body['docs']) == 1
-        assert body['docs'][0]['_source'] == (
-            OPENSEARCH_INGEST_PIPELINE_TEST_CONFIG_1.input_document
-        )
-
-
-class TestGetOpenSearchIngestPipelineTestActualDocumentsFromSimulateResponse:
-    def test_should_extract_actual_documents_from_source(self):
-        actual_documents = (
-            get_opensearch_ingest_pipeline_test_actual_documents_from_simulate_response({
-                'docs': [{
-                    'doc': {
-                        '_source': OPENSEARCH_INGEST_PIPELINE_TEST_CONFIG_1.expected_document
-                    }
-                }]
-            })
-        )
-        assert actual_documents == [OPENSEARCH_INGEST_PIPELINE_TEST_CONFIG_1.expected_document]
-
-
-class TestAssertOpenSearchIngestPipelineTestActualDocumentsMatchExpected:
-    def test_should_pass_if_actual_documents_match_expected_documents(self):
-        assert_opensearch_ingest_pipeline_test_actual_documents_match_expected(
-            actual_documents=[OPENSEARCH_INGEST_PIPELINE_TEST_CONFIG_1.expected_document],
-            ingest_pipeline_config=dataclasses.replace(
-                OPENSEARCH_INGEST_PIPELINE_CONFIG_1,
-                tests=[OPENSEARCH_INGEST_PIPELINE_TEST_CONFIG_1]
-            )
-        )
-
-    def test_should_raise_assertion_error_for_not_matching_documents(self):
-        with pytest.raises(AssertionError):
-            assert_opensearch_ingest_pipeline_test_actual_documents_match_expected(
-                actual_documents=[{'name': 'not matching actual document'}],
-                ingest_pipeline_config=dataclasses.replace(
-                    OPENSEARCH_INGEST_PIPELINE_CONFIG_1,
-                    tests=[OPENSEARCH_INGEST_PIPELINE_TEST_CONFIG_1]
-                )
-            )
-
-
-class TestRunOpenSearchIngestPipelineTests:
-    def test_should_simulate_and_assert_documents_match(
-        self,
-        opensearch_client_mock: MagicMock,
-        assert_opensearch_ingest_pipeline_test_actual_documents_match_expected_mock: MagicMock
-    ):
-        ingest_pipeline_config = dataclasses.replace(
-            OPENSEARCH_INGEST_PIPELINE_CONFIG_1,
-            tests=[OPENSEARCH_INGEST_PIPELINE_TEST_CONFIG_1]
-        )
-        opensearch_client_mock.ingest.simulate.return_value = {
-            'docs': [{
-                'doc': {
-                    '_source': {'name': 'actual document 1'}
-                }
-            }]
-        }
-        run_opensearch_ingest_pipeline_tests(
-            client=opensearch_client_mock,
-            ingest_pipeline_config=ingest_pipeline_config
-        )
-        opensearch_client_mock.ingest.simulate.assert_called_with(
-            body=get_opensearch_ingest_pipeline_tests_body(
-                ingest_pipeline_config
-            )
-        )
-        (
-            assert_opensearch_ingest_pipeline_test_actual_documents_match_expected_mock
-            .assert_called_with(
-                actual_documents=[{'name': 'actual document 1'}],
-                ingest_pipeline_config=ingest_pipeline_config
-            )
-        )
-
-
-class TestCreateOrUpdateOpenSearchIngestPipeline:
-    def test_should_put_ingest_pipeline(
-        self,
-        opensearch_client_mock: MagicMock
-    ):
-        create_or_update_opensearch_ingest_pipeline(
-            client=opensearch_client_mock,
-            ingest_pipeline_config=OPENSEARCH_INGEST_PIPELINE_CONFIG_1
-        )
-        opensearch_client_mock.ingest.put_pipeline.assert_called_with(
-            id=OPENSEARCH_INGEST_PIPELINE_CONFIG_1.name,
-            body=OPENSEARCH_INGEST_PIPELINE_CONFIG_1.definition
-        )
-
-    def test_should_run_ingest_pipeline_tests_if_defined(
-        self,
-        opensearch_client_mock: MagicMock,
-        run_opensearch_ingest_pipeline_tests_mock: MagicMock
-    ):
-        ingest_pipeline_config = dataclasses.replace(
-            OPENSEARCH_INGEST_PIPELINE_CONFIG_1,
-            tests=[OPENSEARCH_INGEST_PIPELINE_TEST_CONFIG_1]
-        )
-        create_or_update_opensearch_ingest_pipeline(
-            client=opensearch_client_mock,
-            ingest_pipeline_config=ingest_pipeline_config
-        )
-        run_opensearch_ingest_pipeline_tests_mock.assert_called_with(
-            client=opensearch_client_mock,
-            ingest_pipeline_config=ingest_pipeline_config
-        )
-
-
 class TestCreateOrUpdateOpenSearchIndex:
     def test_should_create_index_without_settings_if_it_does_not_exist(
         self,
@@ -648,7 +501,7 @@ class TestSetupOpenSearch:
     def test_should_pass_config_to_create_or_update_opensearch_ingestion_pipeline_method(
         self,
         opensearch_client_mock: MagicMock,
-        create_or_update_opensearch_ingest_pipeline_mock: MagicMock
+        create_or_update_opensearch_ingest_pipelines_mock: MagicMock
     ):
         setup_opensearch(
             client=opensearch_client_mock,
@@ -657,9 +510,9 @@ class TestSetupOpenSearch:
                 ingest_pipelines=[OPENSEARCH_INGEST_PIPELINE_CONFIG_1]
             )
         )
-        create_or_update_opensearch_ingest_pipeline_mock.assert_called_with(
+        create_or_update_opensearch_ingest_pipelines_mock.assert_called_with(
             client=opensearch_client_mock,
-            ingest_pipeline_config=OPENSEARCH_INGEST_PIPELINE_CONFIG_1
+            ingest_pipeline_config_list=[OPENSEARCH_INGEST_PIPELINE_CONFIG_1]
         )
 
     def test_should_pass_config_to_create_or_update_opensearch_index_method(
